requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
ansible==7.7.0
ansible-core==2.14.17
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
Brlapi==0.8.2
certifi==2024.8.30
cffi==1.14.5
chardet==4.0.0
click==8.1.7
cockpit==323
cryptography==36.0.1
cupshelpers==1.0
dasbus==1.4
dbus-python==1.2.18
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
file-magic==0.4.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
frozenlist==1.4.1
gpg==1.15.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==2.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-core==0.3.6
langchain-text-splitters==0.3.0
langsmith==0.1.129
libcomps==0.1.18
lxml==4.6.5
MarkupSafe==2.1.5
multidict==6.1.0
nftables==0.1
numpy==1.26.4
openai==1.50.2
orjson==3.10.7
packaging==24.1
perf==0.1
pexpect==4.8.0
pip==21.3.1
ply==3.11
psutil==5.8.0
ptyprocess==0.6.0
pycairo==1.20.1
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
pydantic==2.9.2
pydantic_core==2.23.4
PyGObject==3.40.1
pymongo==4.9.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==1.0.1
python-linux-procfs==0.7.3
pyudev==0.22.0
PyYAML==5.4.1
requests==2.25.1
resolvelib==0.5.4
rpm==4.16.1.3
selinux==3.6
sepolicy==3.6
setools==4.4.4
setroubleshoot==3.3.32
setuptools==53.0.0
six==1.15.0
sniffio==1.3.1
sos==4.7.2
SQLAlchemy==2.0.35
systemd-python==234
tenacity==8.5.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==1.26.5
Werkzeug==3.0.4
yarl==1.13.1
zipp==3.20.2
